Refactor Tag model into Person package
