+++ src/main/java/networkbook/logic/commands/EditCommand.java
@@ -23,7 +23,7 @@ import networkbook.model.person.Person;
 import networkbook.model.person.Phone;
 import networkbook.model.person.Priority;
 import networkbook.model.person.Specialisation;
-import networkbook.model.tag.Tag;
+import networkbook.model.person.Tag;
 import networkbook.model.util.UniqueList;
 
 /**
