+++ src/main/java/networkbook/logic/parser/EditCommandParser.java
@@ -16,7 +16,7 @@ import networkbook.model.person.Email;
 import networkbook.model.person.Link;
 import networkbook.model.person.Phone;
 import networkbook.model.person.Specialisation;
-import networkbook.model.tag.Tag;
+import networkbook.model.person.Tag;
 import networkbook.model.util.UniqueList;
 
 /**
