+++ src/main/java/networkbook/model/util/SampleDataUtil.java
@@ -13,7 +13,7 @@ import networkbook.model.person.Person;
 import networkbook.model.person.Phone;
 import networkbook.model.person.Priority;
 import networkbook.model.person.Specialisation;
-import networkbook.model.tag.Tag;
+import networkbook.model.person.Tag;
 
 /**
  * Contains utility methods for populating {@code NetworkBook} with sample data.
