+++ src/test/java/networkbook/logic/parser/AddCommandParserTest.java
@@ -18,7 +18,7 @@ import networkbook.model.person.Name;
 import networkbook.model.person.Phone;
 import networkbook.model.person.Priority;
 import networkbook.model.person.Specialisation;
-import networkbook.model.tag.Tag;
+import networkbook.model.person.Tag;
 import networkbook.testutil.EditPersonDescriptorBuilder;
 import networkbook.testutil.TypicalIndexes;
