+++ src/test/java/networkbook/logic/parser/CreateCommandParserTest.java
@@ -23,7 +23,7 @@ import networkbook.model.person.Name;
 import networkbook.model.person.Person;
 import networkbook.model.person.Phone;
 import networkbook.model.person.Specialisation;
-import networkbook.model.tag.Tag;
+import networkbook.model.person.Tag;
 import networkbook.testutil.PersonBuilder;
 import networkbook.testutil.TypicalPersons;
