+++ src/test/java/networkbook/logic/parser/EditCommandParserTest.java
@@ -16,7 +16,7 @@ import networkbook.model.person.Link;
 import networkbook.model.person.Name;
 import networkbook.model.person.Phone;
 import networkbook.model.person.Specialisation;
-import networkbook.model.tag.Tag;
+import networkbook.model.person.Tag;
 import networkbook.testutil.EditPersonDescriptorBuilder;
 import networkbook.testutil.TypicalIndexes;
