+++ src/test/java/networkbook/testutil/PersonBuilder.java
@@ -12,7 +12,7 @@ import networkbook.model.person.Person;
 import networkbook.model.person.Phone;
 import networkbook.model.person.Priority;
 import networkbook.model.person.Specialisation;
-import networkbook.model.tag.Tag;
+import networkbook.model.person.Tag;
 import networkbook.model.util.SampleDataUtil;
 import networkbook.model.util.UniqueList;
